refactor(settings): move existing settings to new Performance section
