+++ lib/screens/settings/app_settings/advanced_settings/auto_refresh_time_screen.dart
